pyproject.toml
[project]
name = "iachatconverses"
version = "0.1.0"
description = ""
authors = [
    {name = "Eric Clota Ortiz",email = "eclota@eio.cat"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "langchain (>=0.3.20,<0.4.0)",
    "langgraph (>=0.3.10,<0.4.0)",
    "langchain-google-genai (>=2.1.0,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
]
[tool.poetry]
packages = [{include = "iachatconverses", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
